Update for async/await
The async/await syntax requires to use the 2018 edition, which removes the need
for `extern crate` definitions (which this patch removes).

The API is left untouched, with the exception that `ProxyStream` no longer
implements `Read` and `Write` (`AsyncRead` and `AsyncWrite` no longer require
`Read` and `Write` respectively).

The implementation of `Connect` for `ProxyConnector` is a bit unfortunate in
that it mixes both 'old' future adapter style and the `async` style.  For the
'tls' case this is requires, as otherwise the lifetimes for both `tls` and
`host` do not work out in the close using the `TlsConnector`.  In the 'rustls'
case using the `async` syntax removes the need to create an owned `DNSName` from
the `DNSNameRef` struct.  Likewise we can't just use the `async` syntax, because
the `connect` method takes a plain shared reference to `self`, so we can't have
any references to `self` in any futures we create.